pyproject.toml
[build-system]
requires = ["setuptools>=63", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safespeak"
version = "0.1.0"
description = "Multilingual toxicity moderation research stack for SafeSpeak"
authors = [{ name = "SafeSpeak Team" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "emoji>=2.6.0",
    "unidecode>=1.3.6",
    "pyyaml>=6.0",
    "tqdm>=4.66.0"
]

[project.scripts]
safespeak-ingest = "safespeak.data.ingest:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.1.0"
]
translation = [
    "transformers>=4.40.0",
    "torch>=2.1.0",
    "sentencepiece>=0.1.99",
    "sacremoses>=0.0.53"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
